backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.1
requests==2.32.4
numpy==2.2.6
pandas==2.3.1
scikit-learn==1.7.1
xgboost==3.0.4
torch==2.8.0
torchvision==0.23.0
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
Pillow==11.3.0
tqdm==4.67.1
matplotlib==3.10.5
seaborn==0.13.2
